requirements.txt
aniso8601==9.0.1
click==8.1.3
colorama==0.4.4
Flask==2.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.29
mysqlclient==2.1.0
protobuf==4.21.0
PyJWT==2.4.0
PyMySQL==1.0.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.36
Werkzeug==2.1.2
